requirements.txt
openai>=1.0
python-gitlab>=3.10
GitPython>=3.1

